HealthNet/requirements.txt
blessings==1.6
curtsies==0.2.9
Django==1.10.1
django-bootstrap==0.2.4
django-bootstrap3==7.1.0
django-bootstrap3-datepicker==0.4.1
django-bootstrap3-datetimepicker==2.2.3
greenlet==0.4.10
Pygments==2.1.3
requests==2.11.1
six==1.10.0

wcwidth==0.1.7
